+++ hash_wrappers.cpp
@@ -1,6 +1,5 @@
 #include <array>
 #include <random>
-#include <iostream>
 
 #include "hash_wrappers.h"
 #include "hashes/cityhash/city.h"
@@ -11,9 +10,14 @@
 #include "hashes/murmur_hash/MurmurHash2.h"
 #include "hashes/pearson/pearson.h"
 #include "hashes/super_fast_hash/super_fast_hash.h"
-#include "hashes/various_hash_functions.h"
+#include "hashes/other_hash_functions.h"
 #include "hashes/t1ha/t1ha.h"
 #include "hashes/metro_hash/metrohash64.h"
+#include "hashes/fasthash/fasthash.h"
+#include "hashes/fnv_hash_32/const_fvn_hash.h"
+#include "hashes/fnv_hash_64_128/fnv.h"
+#include "hashes/spooky_hash/spooky.h"
+#include "hashes/jenkins_hash/lookup3.h"
 
 
 namespace hash_wrappers {
@@ -35,21 +39,21 @@ namespace hash_wrappers {
 //----------- CityHashes ----------
 
     size_t CityHash32Wrapper::operator()(const std::string &str) const {
-        return CityHash32(str.c_str(), str.size());
+        return city::CityHash32(str.c_str(), str.size());
     }
 
     size_t CityHash64Wrapper::operator()(const std::string &str) const {
-        return CityHash64(str.c_str(), str.size());
+        return city::CityHash64(str.c_str(), str.size());
     }
 
     size_t CityHash64WithSeedWrapper::operator()(const std::string &str) const {
-        return CityHash64WithSeed(str.c_str(), str.size(), GenerateSeed());
+        return city::CityHash64WithSeed(str.c_str(), str.size(), GenerateSeed());
     }
 
 
     size_t CityHash64WithSeedsWrapper::operator()(const std::string &str) const {
         const auto[seed0, seed1] = GenerateTwoSeeds();
-        return CityHash64WithSeeds(str.c_str(), str.size(), seed0, seed1);
+        return city::CityHash64WithSeeds(str.c_str(), str.size(), seed0, seed1);
     }
 
 //----------- FarmHashes ----------
@@ -97,59 +101,48 @@ namespace hash_wrappers {
 //---------- MurmurHashes ---------
 
     size_t MurmurHash1Wrapper::operator()(const std::string &str) const {
-        return MurmurHash1(str.c_str(), str.size(), GenerateSeed());
+        return MurmurHash1(str.c_str(), static_cast<int>(str.size()), GenerateSeed());
     }
 
     size_t MurmurHash1AlignedWrapper::operator()(const std::string &str) const {
-        return MurmurHash1Aligned(str.c_str(), str.size(), GenerateSeed());
+        return MurmurHash1Aligned(str.c_str(), static_cast<int>(str.size()), GenerateSeed());
     }
 
     size_t MurmurHash2Wrapper::operator()(const std::string &str) const {
-        return MurmurHash2(str.c_str(), str.size(), GenerateSeed());
+        return MurmurHash2(str.c_str(), static_cast<int>(str.size()), GenerateSeed());
     }
 
     size_t MurmurHash64AWrapper::operator()(const std::string &str) const {
-        return MurmurHash64A(str.c_str(), str.size(), GenerateSeed());
+        return MurmurHash64A(str.c_str(), static_cast<int>(str.size()), GenerateSeed());
     }
 
     size_t MurmurHash64BWrapper::operator()(const std::string &str) const {
-        return MurmurHash64B(str.c_str(), str.size(), GenerateSeed());
+        return MurmurHash64B(str.c_str(), static_cast<int>(str.size()), GenerateSeed());
     }
 
 //---------- PearsonHashes ---------
 
     size_t PearsonHash32Wrapper::operator()(const std::string &str) const {
-        void* in = (void*)str.data();
         return pearson_hash_32((const unsigned char*)str.c_str(), str.size(), 0);
     }
 
     size_t PearsonHash64Wrapper::operator()(const std::string &str) const {
-
-        void* in = (void*)str.data();
-        return pearson_hash_64((uint8_t*)in, str.size(), 0);
+        return pearson_hash_64((uint8_t*)str.c_str(), str.size(), 0);
     }
 
 //--- Paul Hsieh's SuperFastHash ---
 
     size_t SuperFastHashWrapper::operator()(const std::string &str) const {
-        return SuperFastHash(str.data(), str.size());
+        return SuperFastHash(str.data(), static_cast<int>(str.size()));
     }
 
 //----- Rolling Hash (BuzHash) -----
 
-    size_t CyclicHash32Wrapper::operator()(const std::string &str) const {
+    size_t Buzhash32Wrapper::operator()(const std::string &str) const {
         return hasher_.hash(str);
     }
 
-    size_t CyclicHash64Wrapper::operator()(const std::string &str) const {
-        return hasher_.hash(str);
-    }
-
-    size_t RabinKarpHash32Wrapper::operator()(const std::string &str) const {
-        return hasher_.hash(str);
-    }
-
-    size_t RabinKarpHash64Wrapper::operator()(const std::string &str) const {
+    size_t Buzhash64Wrapper::operator()(const std::string &str) const {
         return hasher_.hash(str);
     }
 
@@ -206,6 +199,71 @@ namespace hash_wrappers {
         return MetroHashConvert(metrohash64_2, str);
     }
 
+
+//------------ FastHash ------------
+
+    size_t FastHash32Wrapper::operator()(const std::string &str) const {
+        return fasthash32(str.data(), str.size(), GenerateSeed());
+    }
+
+    size_t FastHash64Wrapper::operator()(const std::string &str) const {
+        return fasthash64(str.data(), str.size(), GenerateSeed());
+    }
+
+//----- Bernstein's hash djb2 ------
+
+    size_t DJB2Hash32Wrapper::operator()(const std::string &str) const {
+        return DJB2Hash32(str);
+    }
+
+    size_t DJB2Hash64Wrapper::operator()(const std::string &str) const {
+        return DJB2Hash64(str);
+    }
+
+//------------ PJW Hash ------------
+
+    size_t PJWHash32Wrapper::operator()(const std::string &str) const {
+        return PJWHash32(str);
+    }
+
+    size_t PJWHash64Wrapper::operator()(const std::string &str) const {
+        return PJWHash64(str);
+    }
+
+//---------- FNV-1a hash -----------
+
+    size_t FNV_1a_32Wrapper::operator()(const std::string &str) const {
+        return hardrock::FnvHash(str.data());
+    }
+
+    size_t FNV_1a_64Wrapper::operator()(const std::string &str) const {
+        return net::QuicUtils::FNV1a_64_Hash(str.data(), static_cast<int>(str.size()));
+    }
+
+//---------- Spooky hash -----------
+
+    size_t SpookyHash32Wrapper::operator()(const std::string &str) const {
+        return spooky_hash32(str.data(), str.size(), GenerateSeed());
+    }
+
+    size_t SpookyHash64Wrapper::operator()(const std::string &str) const {
+        return spooky_hash64(str.data(), str.size(), GenerateSeed());
+    }
+
+    //--------- Jenkins hash -----------
+
+    uint32_t JenkinsOneTimeHashWrapper::operator()(const std::string &str) const {
+        return jenkins_one_at_a_time_hash((uint8_t*)str.data(), str.size());
+    }
+
+    uint32_t Lookup3LittleWrapper::operator()(const std::string &str) const {
+        return hashlittle(str.data(), str.size(), GenerateSeed());
+    }
+
+    uint32_t Lookup3BigWrapper::operator()(const std::string &str) const {
+        return hashbig(str.data(), str.size(), GenerateSeed());
+    }
+
 //----------- BuildHashes ----------
 
     std::vector<Hasher> Build32bitsHashes() {
@@ -229,13 +287,26 @@ namespace hash_wrappers {
 
         hashes.push_back({"Paul Hsieh's SuperFastHash", SuperFastHashWrapper{}});
 
-        hashes.push_back({"RollingHash (CyclicHash32)", CyclicHash32Wrapper{}});
-        //hashes.push_back({"RollingHash (RabinKarpHash32)", RabinKarpHash32Wrapper{}});
+        hashes.push_back({"RollingHash (BuzHash32)", Buzhash32Wrapper{}});
 
         hashes.push_back({"SDBMHash32", SDBMHash32Wrapper{}});
 
         hashes.push_back({"t1ha0_32le", t1ha0_32le_wrapper{}});
         hashes.push_back({"t1ha0_32be", t1ha0_32be_wrapper{}});
+
+        hashes.push_back({"fasthash32", FastHash32Wrapper{}});
+
+        hashes.push_back({"Bernstein's hash djb2 32", DJB2Hash32Wrapper{}});
+
+        hashes.push_back({"PJW Hash 32", PJWHash32Wrapper{}});
+
+        hashes.push_back({"FNV-1a hash 32", FNV_1a_32Wrapper{}});
+
+        hashes.push_back({"Spooky hash 32", SpookyHash32Wrapper{}});
+
+        hashes.push_back({"Jenkins hash (One at a time)", JenkinsOneTimeHashWrapper{}});
+        hashes.push_back({"Jenkins hash (lookup3 - little)", Lookup3LittleWrapper{}});
+        hashes.push_back({"Jenkins hash (lookup3 - big)", Lookup3BigWrapper{}});
         return hashes;
     }
 
@@ -261,8 +332,7 @@ namespace hash_wrappers {
         pearson_hash_init();
         hashes.push_back({"PearsonHash64", PearsonHash64Wrapper{}});
 
-        hashes.push_back({"RollingHash (CyclicHash64)", CyclicHash64Wrapper{}});
-        //hashes.push_back({"RollingHash (RabinKarpHash64)", RabinKarpHash64Wrapper{}});
+        hashes.push_back({"RollingHash (BuzHash64)", Buzhash64Wrapper{}});
 
         hashes.push_back({"SDBMHash64", SDBMHash64Wrapper{}});
 
@@ -271,7 +341,17 @@ namespace hash_wrappers {
         hashes.push_back({"t1ha2_atonce", t1ha2_atonce_wrapper{}});
 
         hashes.push_back({"metrohash64_1", MetroHash64_1_Wrapper{}});
-        hashes.push_back({"metrohash64_1", MetroHash64_2_Wrapper{}});
+        hashes.push_back({"metrohash64_2", MetroHash64_2_Wrapper{}});
+
+        hashes.push_back({"fasthash64", FastHash64Wrapper{}});
+
+        hashes.push_back({"Bernstein's hash djb2 64", DJB2Hash64Wrapper{}});
+
+        hashes.push_back({"PJW Hash 64", PJWHash64Wrapper{}});
+
+        hashes.push_back({"FNV-1a hash 64", FNV_1a_64Wrapper{}});
+
+        hashes.push_back({"Spooky hash 64", SpookyHash64Wrapper{}});
         return hashes;
     }
 }