+++ changes.md
@@ -5,3 +5,7 @@
 
 Изменения 06.04.2021:
 1. Добавил SDBM hash, t1ha hash (заработало со второй попытки), MetroHash
+
+Изменения 08.04.2021:
+1. Добавил Fast-Hash, Bernstein's hash djb2, FNV-1a hash, Spooky hash, Jenkins hash (one_at_a_time, lookup3)
+2. Включил хеш-функции CityHash в namespace city, так как бл конфликто типов uint128